refactor: add App module
